chat-backend/requirements.txt
supabase==2.15.2
requests==2.32.3
dotenv==0.9.9
fastapi==0.115.12
uvicorn==0.34.3
pydantic-settings==2.9.1
python-jose[cryptography]==3.3.0
cryptography==43.0.1
python-multipart==0.0.9